Factored out nodey sdi driver
